refactor: Move catalog pages under their own folder for clarity
