+++ app/docs/catalog/chart-config.mdx
