+++ app/docs/catalog/chart-config.tsx
