+++ app/docs/catalog/index.mdx
@@ -1,5 +1,3 @@
-import ImageSpecimen from './Image'
-
 > Documentation for the Linked Data «Basistool Visualisierung».
 
 <ImageSpecimen src="./static/docs/hero.png" plain />
