+++ app/docs/catalog/layout.mdx
@@ -1,4 +1,3 @@
-import TableSpecimen from './TableSpecimen'
 
 ### Spacings
