+++ app/docs/chart-selection-tabs.docs.mdx
@@ -5,7 +5,6 @@ import {
 } from "@/configurator";
 import palmerPenguinsFixture from "@/test/__fixtures/config/int/scatterplot-palmer-penguins.json";
 import { Canvas, Meta } from '@storybook/blocks';
-import { ReactSpecimen } from './catalog';
 import * as ChartSelectionTabsStories from './chart-selection-tabs.stories';
 
 > Chart Selection Tabs are used in ChartPreview and ChartPublished to enable selecting a chart type in any stage of creating a chart.
@@ -22,7 +21,6 @@ They can be either _editable_, to display a button to show ChartTypeSelector (us
 
 {/* ## Non-editable
 
-<ReactSpecimen span={6}>
   <ConfiguratorStateProvider
     chartId={palmerPenguinsFixture.key}
     initialState={
@@ -35,7 +33,7 @@ They can be either _editable_, to display a button to show ChartTypeSelector (us
   >
     <ChartSelectionTabs />
   </ConfiguratorStateProvider>
-</ReactSpecimen> */}
+ */}
 
 ## For developers
