+++ app/pages/docs/index.tsx
@@ -26,17 +26,17 @@ const pages: ConfigPageOrGroup[] = [
   {
     path: "/",
     title: "Introduction",
-    content: require("@/docs/index.mdx"),
+    content: require("@/docs/catalog/index.mdx"),
   },
   {
     path: "/branding",
     title: "Branding",
-    content: require("@/docs/branding.mdx"),
+    content: require("@/docs/catalog/branding.mdx"),
   },
   {
     path: "/accessibility",
     title: "Accessibility",
-    content: require("@/docs/accessibility.mdx"),
+    content: require("@/docs/catalog/accessibility.mdx"),
   },
   {
     title: "Theming",
@@ -44,27 +44,17 @@ const pages: ConfigPageOrGroup[] = [
       {
         path: "/theming",
         title: "Overview",
-        content: require("@/docs/theming.mdx"),
+        content: require("@/docs/catalog/theming.mdx"),
       },
       {
         path: "/colors",
         title: "Colors",
-        content: require("@/docs/colors.mdx"),
-      },
-      {
-        path: "/typography",
-        title: "Typography",
-        content: require("@/docs/typography.mdx"),
-      },
-      {
-        path: "/iconography",
-        title: "Iconography",
-        content: require("@/docs/icons.docs.tsx"),
+        content: require("@/docs/catalog/colors.mdx"),
       },
       {
         path: "/layout",
         title: "Layout",
-        content: require("@/docs/layout.mdx"),
+        content: require("@/docs/catalog/layout.mdx"),
       },
     ],
   },
@@ -74,12 +64,12 @@ const pages: ConfigPageOrGroup[] = [
       {
         path: "/mockups",
         title: "Mockups",
-        content: require("@/docs/mockups"),
+        content: require("@/docs/catalog/mockups"),
       },
       {
         path: "/chart-config",
         title: "Chart-Config",
-        content: require("@/docs/chart-config"),
+        content: require("@/docs/catalog/chart-config"),
       },
     ],
   },
@@ -89,37 +79,12 @@ const pages: ConfigPageOrGroup[] = [
       {
         path: "/charts/rdf-to-visualize",
         title: "RDF to visualize",
-        content: require("@/docs/rdf-to-visualize.mdx"),
+        content: require("@/docs/catalog/rdf-to-visualize.mdx"),
       },
       {
         path: "/charts/preview-via-api",
         title: "Preview via API",
-        content: require("@/docs/chart-preview-via-api.mdx"),
-      },
-      {
-        path: "/charts/annotations",
-        title: "Annotations",
-        content: require("@/docs/annotations.docs"),
-      },
-      {
-        path: "/charts/columns-chart",
-        title: "Columns",
-        content: require("@/docs/columns.docs"),
-      },
-      {
-        path: "/charts/line-chart",
-        title: "Lines",
-        content: require("@/docs/lines.docs"),
-      },
-      {
-        path: "/charts/scatterplot",
-        title: "Scatterplot",
-        content: require("@/docs/scatterplot.docs"),
-      },
-      {
-        path: "/charts/data-table",
-        title: "Table",
-        content: require("@/docs/data-table.docs"),
+        content: require("@/docs/catalog/chart-preview-via-api.mdx"),
       },
     ],
   },
@@ -129,94 +94,14 @@ const pages: ConfigPageOrGroup[] = [
       {
         path: "/components/intro",
         title: "Introduction",
-        content: require("@/docs/components"),
-      },
-      {
-        path: "/components/button",
-        title: "Button",
-        content: require("@/docs/button.docs"),
-      },
-      {
-        path: "/components/chart-selection-tabs",
-        title: "Chart Selection Tabs",
-        content: require("@/docs/chart-selection-tabs.docs"),
-      },
-      {
-        path: "/components/color-ramp",
-        title: "Color Ramp",
-        content: require("@/docs/color-ramp.docs"),
-      },
-      {
-        path: "/components/controls",
-        title: "Controls",
-        content: require("@/docs/controls.docs"),
-      },
-      {
-        path: "/components/dataset-browse",
-        title: "Datasets Browsing",
-        content: require("@/docs/dataset-browse.docs"),
-      },
-      {
-        path: "/components/footer",
-        title: "Footer",
-        content: require("@/docs/footer.docs"),
-      },
-      {
-        path: "/components/form",
-        title: "Form",
-        content: require("@/docs/form.docs"),
-      },
-      {
-        path: "/components/header",
-        title: "Header",
-        content: require("@/docs/header.docs"),
-      },
-      {
-        path: "/components/hints",
-        title: "Hints",
-        content: require("@/docs/hint.docs"),
-      },
-      {
-        path: "/components/homepage",
-        title: "Homepage",
-        content: require("@/docs/homepage.docs"),
-      },
-      {
-        path: "/components/publish-actions",
-        title: "Publish actions",
-        content: require("@/docs/publish-actions.docs"),
-      },
-      {
-        path: "/components/table",
-        title: "Preview Table",
-        content: require("@/docs/table.docs"),
-      },
-      {
-        path: "/components/tags",
-        title: "Tags",
-        content: require("@/docs/tags.docs"),
-      },
-      {
-        path: "/components/tabs",
-        title: "Tabs",
-        content: require("@/docs/tabs.docs"),
-      },
-      {
-        path: "/components/tooltips",
-        title: "Tooltips",
-        content: require("@/docs/tooltips.docs"),
-      },
-      {
-        path: "/components/typography",
-        title: "Typography",
-        content: require("@/docs/text.docs"),
+        content: require("@/docs/catalog/components"),
       },
     ],
   },
   {
     path: "/testing",
     title: "Testing",
-    content: require("@/docs/testing.mdx"),
+    content: require("@/docs/catalog/testing.mdx"),
   },
 ];
 const mkHeading = (level: number) => {
